+++ __test__/specs/props.spec.js
@@ -3,15 +3,11 @@ import {shallow} from 'vue-test-utils'
 import Component from '../../src/component.vue';
 import moment from 'moment';
 
-describe('Datepicker props', () => {
+describe('datepicker props', () => {
 
   // Store for future usage
   const props = {
     value: '10/11/2017',
-    inputClass: 'input-field',
-    placeholder: 'select-a-date',
-    name: 'dob',
-    id: 'dob-field',
     config: {
       format: 'DD/MM/YYYY'
     }
@@ -28,26 +24,6 @@ describe('Datepicker props', () => {
 
   afterEach(() => {
     wrapper.vm.$destroy();
-  });
-
-  test('accepts css class via prop', () => {
-    expect(wrapper.hasProp('inputClass', props.inputClass)).toBe(true);
-    expect(wrapper.hasClass(props.inputClass)).toBe(true);
-  });
-
-  test('accepts input name via prop', () => {
-    expect(wrapper.hasProp('name', props.name)).toBe(true);
-    expect(wrapper.hasAttribute('name', props.name)).toBe(true);
-  });
-
-  test('accepts input id via prop', () => {
-    expect(wrapper.hasProp('id', props.id)).toBe(true);
-    expect(wrapper.hasAttribute('id', props.id)).toBe(true);
-  });
-
-  test('accepts input placeholder via prop', () => {
-    expect(wrapper.hasProp('placeholder', props.placeholder)).toBe(true);
-    expect(wrapper.hasAttribute('placeholder', props.placeholder)).toBe(true);
   });
 
   test('accepts config via prop', () => {